--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -11,42 +11,43 @@ jobs:
   linux-unit-tests:
     needs: yarn-lockfile-check
     uses: salesforcecli/github-workflows/.github/workflows/unitTestsLinux.yml@main
-  windows-unit-tests:
-    needs: linux-unit-tests
-    uses: salesforcecli/github-workflows/.github/workflows/unitTestsWindows.yml@main
+  # windows-unit-tests:
+  #   needs: linux-unit-tests
+  #   uses: salesforcecli/github-workflows/.github/workflows/unitTestsWindows.yml@main
   nuts:
     needs: linux-unit-tests
     uses: salesforcecli/github-workflows/.github/workflows/nut.yml@main
     secrets: inherit
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest]
+        os: [windows-latest]
       fail-fast: false
     with:
       os: ${{ matrix.os }}
-  xNuts-ps:
-    needs: linux-unit-tests
-    uses: salesforcecli/github-workflows/.github/workflows/externalNut.yml@main
-    strategy:
-      fail-fast: false
-      matrix:
-        os: ['ubuntu-latest', 'windows-latest']
-        command:
-          - 'yarn test:nuts:tracking:basics'
-          - 'yarn test:nuts:tracking:conflicts'
-          - 'yarn test:nuts:tracking:forceignore'
-          - 'yarn test:nuts:tracking:remote'
-          - 'yarn test:nuts:tracking:lwc'
-          - 'yarn test:nuts:deb'
-          - 'yarn mocha "test/nuts/trackingCommands/mpd-*.nut.ts" --slow 3000 --timeout 600000 --parallel'
-    with:
-      packageName: '@salesforce/source-tracking'
-      externalProjectGitUrl: 'https://github.com/salesforcecli/plugin-source'
-      command: ${{matrix.command}}
-      os: ${{matrix.os}}
-      preSwapCommands: 'yarn upgrade @salesforce/source-deploy-retrieve@latest; npx yarn-deduplicate; yarn install'
-      useCache: false
-    secrets: inherit
+      command: 'git config core.autocrlf true; git config --list; yarn test:nuts'
+  # xNuts-ps:
+  #   needs: linux-unit-tests
+  #   uses: salesforcecli/github-workflows/.github/workflows/externalNut.yml@main
+  #   strategy:
+  #     fail-fast: false
+  #     matrix:
+  #       os: ['ubuntu-latest', 'windows-latest']
+  #       command:
+  #         - 'yarn test:nuts:tracking:basics'
+  #         - 'yarn test:nuts:tracking:conflicts'
+  #         - 'yarn test:nuts:tracking:forceignore'
+  #         - 'yarn test:nuts:tracking:remote'
+  #         - 'yarn test:nuts:tracking:lwc'
+  #         - 'yarn test:nuts:deb'
+  #         - 'yarn mocha "test/nuts/trackingCommands/mpd-*.nut.ts" --slow 3000 --timeout 600000 --parallel'
+  #   with:
+  #     packageName: '@salesforce/source-tracking'
+  #     externalProjectGitUrl: 'https://github.com/salesforcecli/plugin-source'
+  #     command: ${{matrix.command}}
+  #     os: ${{matrix.os}}
+  #     preSwapCommands: 'yarn upgrade @salesforce/source-deploy-retrieve@latest; npx yarn-deduplicate; yarn install'
+  #     useCache: false
+  #   secrets: inherit
 
   xNuts-pdr:
     needs: linux-unit-tests
@@ -54,15 +55,14 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ['ubuntu-latest', 'windows-latest']
+        os: ['windows-latest']
         command:
           - 'yarn test:nuts:tracking'
-          - 'yarn test:nuts:deb'
     with:
       packageName: '@salesforce/source-tracking'
       externalProjectGitUrl: 'https://github.com/salesforcecli/plugin-deploy-retrieve'
       command: ${{matrix.command}}
       os: ${{matrix.os}}
-      preSwapCommands: 'yarn upgrade @salesforce/source-deploy-retrieve@latest; npx yarn-deduplicate; yarn install'
+      preSwapCommands: 'git config core.autocrlf true; git config --list; yarn upgrade @salesforce/source-deploy-retrieve@latest; npx yarn-deduplicate; yarn install'
       useCache: false
     secrets: inherit