### What does this PR do?
Testing a suggestion from @jayree to speed up performance on windows

### What issues does this PR fix or reference?
[@W-15983399@](https://gus.my.salesforce.com/apex/ADM_WorkLocator?bugorworknumber=W-15983399)