@iowillhoit 2024-06-17
## Initial findings

* BEFORE
    * Run 1 https://github.com/forcedotcom/source-tracking/actions/runs/9427535056/job/25971785839
        * 2000 moves: 47197ms
        * 200,000 status: 30972ms
        * 200,000 commit: 230575ms
    * Run 2 https://github.com/forcedotcom/source-tracking/actions/runs/9427535056/job/26336899175
        * 2000 moves: 48607ms
        * 200,000 status: 31125ms
        * 200,000 commit: 229347ms
    * Run 3 https://github.com/forcedotcom/source-tracking/actions/runs/9427535056/job/26337557131
        * 2000 moves: 53644ms
        * 200,000 status: 32858ms
        * 200,000 commit: 233199ms
    * Before averages:
        * (47197+48607+53644) / 3 = 49816
        * (30972+31125+32858) / 3 = 31651.66
        * (230575+229347+233199) / 3 = 231040.33

* AFTER
    * Run 1 https://github.com/forcedotcom/source-tracking/actions/runs/9554021573/job/26334245675
        * 2000 moves: 51822ms
        * 200,000 status: 33630ms
        * 200,000 commit: 239251ms
    * Run 2 https://github.com/forcedotcom/source-tracking/actions/runs/9554021573/job/26334245675
        * 2000 moves: 51551ms
        * 200,000 status: 32205ms
        * 200,000 commit: 231066ms
    * Run 3 https://github.com/forcedotcom/source-tracking/actions/runs/9554021573/job/26338275285
        * 2000 moves: 51503ms
        * 200,000 status: 32319ms
        * 200,000 commit: 232452ms
    * After averages:
        * (51822+51551+51503) / 3 = 51625.33
        * (33630+32205+32319) / 3 = 32718
        * (239251+231066+232452) / 3 = 234256.33

* Improvement (was faster before autocrlf change)
    * 49816 / 51625.33 = 0.965
    * 31651.66 / 32718 = 0.967
    * 231040.33 / 234256.33 = 0.986