@iowillhoit 2024-06-19
More testing after setting `core.autocrlf` at the local git config level also. Still no noticeable difference. 
Also looked at the execution time of the xNUTS for PDR and there is also no noticeable difference ([compared to this run without any of the autocrlf changes](https://github.com/forcedotcom/source-tracking/actions/runs/9585412329/job/26431276649?pr=614) 5m 56s)

* AFTER SETTING AUTOCRLF AT THE LOCAL LEVEL
    * Run 1 https://github.com/forcedotcom/source-tracking/actions/runs/9584828592/job/26429352828
        * 2000 moves: 47710ms
        * 200,000 status: 30895ms
        * 200,000 commit: 228553ms
        * xPDR: failed because of deps
    * Run 2 https://github.com/forcedotcom/source-tracking/actions/runs/9585016695/job/26430017268
        * 2000 moves: 50099ms
        * 200,000 status: 32120ms
        * 200,000 commit: 233728ms
        * xPDR: 5m 51s
    * Run 3 https://github.com/forcedotcom/source-tracking/actions/runs/9585016695/job/26431305967
        * 2000 moves: 49051ms
        * 200,000 status: 30532ms
        * 200,000 commit: 226435ms
        * xPRD: 5m 55s